+++ Pipfile
@@ -5,7 +5,7 @@ name = "pypi"
 
 [scripts]
 start-server = "flask run"
-start-monitor = "python -d -s -m monitoring"
+start-monitor = "python -d -s -m monitor"
 start-database = "./scripts/start_database.sh"
 create-package = "tar -cvzf arpi-server.tar.gz --exclude='__pycache__' etc scripts src Pipfile Pipfile.lock"
