+++ src/monitor/adapters/keypads/dsc.py
@@ -4,7 +4,7 @@ from time import sleep, time
 
 import RPi.GPIO as GPIO
 
-from monitoring.adapters.keypads.base import KeypadBase
+from monitor.adapters.keypads.base import KeypadBase
 from constants import LOG_ADKEYPAD
 
 # Magic numbers
