+++ src/monitor/adapters/mock/MCP3008.py
