+++ src/monitor/adapters/mock/output.py
