+++ src/monitor/adapters/smtp.py
