+++ src/monitor/adapters/syren.py
@@ -7,14 +7,14 @@
 import logging
 import os
 
-from monitoring.adapters import SYREN_OUT
+from monitor.adapters import SYREN_OUT
 from constants import LOG_ADSYREN
 
 # check if running on Raspberry
 if os.environ.get("USE_SIMULATOR", "false").lower() == "false":
     from gpiozero.output_devices import DigitalOutputDevice
 else:
-    from monitoring.adapters.mock.output import Output as DigitalOutputDevice
+    from monitor.adapters.mock.output import Output as DigitalOutputDevice
 
 
 class SyrenAdapter(object):
