+++ src/monitor/notifications/templates.py
